Tools/mir4-tools-bot/requirements.txt
discord.py==2.3.1
imutils==0.5.4
numpy
opencv-contrib-python-headless
pytesseract==0.3.10
python-dotenv==1.0.0